+++ lib/stone_bank/operations.ex
@@ -6,6 +6,20 @@ defmodule StoneBank.Operations do
   alias StoneBank.{Accounts, Accounts.Account}
   alias StoneBank.Repo
 
+  def withdraw(f_id, value) do
+    from = Accounts.get!(f_id)
+    value = Money.new(:brl, value)
+
+    case is_negative_amount(from.amount, value) do
+      true ->
+        {:error, "you can't have negative amount!"}
+
+      false ->
+        {:ok, from} = perform_operation(from, value, :sub) |> Repo.update()
+        {:ok, "Withdraw with success!!! from: #{from.id}, value: #{value}"}
+    end
+  end
+
   def transfer(f_id, t_id, value) do
     from = Accounts.get!(f_id)
     value = Money.new(:brl, value)
@@ -22,13 +36,19 @@ defmodule StoneBank.Operations do
   end
 
   defp perform_update(from, t_id, value) do
-    {:ok, from} = perform_operation(from, value, :sub)
+    to = Accounts.get!(t_id)
 
-    {:ok, to} =
-      Accounts.get!(t_id)
-      |> perform_operation(value, :sum)
+    transaction =
+      Ecto.Multi.new()
+      |> Ecto.Multi.update(:account_from, perform_operation(from, value, :sub))
+      |> Ecto.Multi.update(:account_to, perform_operation(to, value, :sum))
+      |> Repo.transaction()
 
-    {:ok, "Transfer with sucess!! from: #{from.id} to: #{to.id} value: #{value}"}
+    case transaction do
+      {:ok, _} -> {:ok, "Transfer with sucess!! from: #{from.id} to: #{to.id} value: #{value}"}
+      {:error, :account_from, changeset, _} -> {:error, changeset}
+      {:error, :account_to, changeset, _} -> {:error, changeset}
+    end
   end
 
   defp perform_operation(account, value, :sub) do
@@ -43,6 +63,5 @@ defmodule StoneBank.Operations do
 
   def update_account(%Account{} = account, params) do
     Account.changeset(account, params)
-    |> Repo.update()
   end
 end