+++ lib/stone_bank_web/controllers/operation_controller.ex
@@ -7,7 +7,7 @@ defmodule StoneBankWeb.OperationController do
   def transfer(conn, %{"from_account_id" => f_id, "to_account_id" => t_id, "value" => value}) do
     with {:ok, message} <- Operations.transfer(f_id, t_id, value) do
       conn
-      |> render("sucess.json", message: message)
+      |> render("success.json", message: message)
     end
   end