+++ lib/stone_bank_web/views/operation_view.ex
@@ -1,7 +1,7 @@
 defmodule StoneBankWeb.OperationView do
   use StoneBankWeb, :view
 
-  def render("sucess.json", %{message: message}) do
+  def render("success.json", %{message: message}) do
     %{
       message: message
     }