+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## Unreleasead
 
-## [0.0.14] - 2021-04-09
+## [0.0.15] - 2021-04-09
 ### Added
 
-- add operations/withdraw route
-- add withdraw def at operation controller
+- add def withdraw at operations module
 
 ### Changed
+
+- refactor(operations) add transaction with Ecto.Multi at transfer function
+- refactor(operation_controller/view) correct the word success to unify both calls
 
 ### Removed